# Conflicts:
#	flickit-assessment-kit/src/test/java/org/flickit/assessment/kit/application/service/assessmentkit/update/impl/QuestionUpdateKitPersisterTest.java
#	flickit-assessment-kit/src/test/java/org/flickit/assessment/kit/test/fixture/application/Constants.java

+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/out/persistence/questionimpact/QuestionImpactPersistenceJpaAdapter.java
@@ -3,6 +3,7 @@ package org.flickit.assessment.kit.adapter.out.persistence.questionimpact;
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.data.jpa.kit.questionimpact.QuestionImpactJpaRepository;
+import org.flickit.assessment.data.jpa.kit.seq.KitDbSequenceGenerators;
 import org.flickit.assessment.kit.application.domain.QuestionImpact;
 import org.flickit.assessment.kit.application.port.out.questionimpact.CreateQuestionImpactPort;
 import org.flickit.assessment.kit.application.port.out.questionimpact.DeleteQuestionImpactPort;
@@ -20,14 +21,20 @@ public class QuestionImpactPersistenceJpaAdapter implements
     UpdateQuestionImpactPort {
 
     private final QuestionImpactJpaRepository repository;
+    private final KitDbSequenceGenerators sequenceGenerators;
 
     @Override
     public Long persist(QuestionImpact impact) {
-        return repository.save(mapToJpaEntityToPersist(impact)).getId();
+        var entity = mapToJpaEntityToPersist(impact);
+        entity.setId(sequenceGenerators.generateQuestionImpactId());
+        return repository.save(entity).getId();
     }
 
     @Override
     public void delete(Long questionImpactId, Long kitVersionId) {
+        if(!repository.existsByIdAndKitVersionId(questionImpactId, kitVersionId))
+            throw new ResourceNotFoundException(QUESTION_IMPACT_ID_NOT_FOUND);
+
         repository.deleteByIdAndKitVersionId(questionImpactId, kitVersionId);
     }
